pyproject.toml
[project]
name = "H3Chat"
version = "0.1.0"
description = "H3 Chat"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.1",
    "langchain-text-splitters>=1.1.1",
    "pgvector>=0.4.2",
    "psycopg2-binary>=2.9.11",
    "pymupdf4llm>=1.27.2.1",
    "sentence-transformers>=5.3.0",
    "tqdm>=4.67.3",
    "python-dotenv>=1.0.0",
    "dotenv>=0.9.9",
    "uvicorn>=0.42.0",
    "python-multipart>=0.0.22",
]
